catalog/scripts/requirements.txt
# Enterprise Observability Platform — Catalog Scripts
# Requirements for OpenMetadata ingestion and governance automation

# OpenMetadata ingestion framework
openmetadata-ingestion[databricks,kafka,clickhouse]==1.3.*

# Data quality
great-expectations==0.18.*
great-expectations-cloud

# Utilities
pydantic>=2.0
python-dotenv>=1.0
tenacity>=8.0          # Retry with backoff
click>=8.0             # CLI helpers
rich>=13.0             # Pretty terminal output
pyyaml>=6.0

# Testing
pytest>=7.4
pytest-cov>=4.1
pytest-mock>=3.12
responses>=0.24        # HTTP mock for API tests
